Fix: round(string) function: locale-dependent code
In some locales, possibly it did not replace "E+" to "E"
of exponent parts of strings storing floating-point numbers
(e.g.: "1.23E+45").

The results are valid as floating-point numbers even if
"E+" are not replaced to "E", but
we don't wnat to make behavior of this function locale-dependent.
So we fixed it.